Directory creation, added log and transcript vs. metadata duration validation
Creates a directory for each conversation that has PROOFREAD or
TRANSCRIBED transcripts. Transcripts marked INVALID, TODO, or
INPROGRESS, as well as test transcripts or transcripts that have none of
the aforementioned tags, will not be added to a directory, but their IDs
and the reason is logged in a text file. Added function to validate
metadata vs. transcript duration.